agentic-engine-server/requirements.txt
langchain==0.1.0
langchain-core==0.1.10
langchain-google-genai==0.0.11
google-generativeai==0.4.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
pytz==2023.3 
fastapi==0.110.0
uvicorn==0.29.0 